requirements.txt
# certifi==2020.12.5
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
cssselect==1.1.0
fake-useragent==0.1.11
idna==2.10
lxml==4.6.1
netaddr==0.8.0
parse==1.18.0
psutil==5.7.3
pyee==7.0.4
pyppeteer==0.2.2
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyquery==1.4.1
python-engineio==3.14.2
python-socketio==4.6.0
requests==2.25.0
requests-html==0.10.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.51.0
urllib3==1.26.1
w3lib==1.22.0
websocket-client==0.57.0
websockets==8.1
webdriver-manager==3.4.1
selenium~=3.141.0